refactor(disponibilidade-professores): remove componentes obsoletos e implementa novo grid de disponibilidade
- Remove os componentes `CreateDisponibilidadeFormDialog`, `EditDisponibilidadeFormDialog`, `DeleteDisponibilidadeAlertDialog`, `DisponibilidadesDataTable`, e `disponibilidadeColumns`, que não são mais utilizados.
- Introduz o componente `GridTurno` para gerenciar a visualização de disponibilidades em turnos, melhorando a interação do usuário.
- Atualiza a página de disponibilidade do professor para integrar o novo grid, permitindo uma configuração mais intuitiva dos horários.
- Refatora o componente `ResumoHorariosCards` para incluir novas funcionalidades e melhorar a apresentação dos dados.
- Adiciona um novo hook `useConfiguracaoHorario` para buscar configurações de horário, facilitando a gestão de slots de disponibilidade.
