+++ src/features/disponibilidade-professores/components/grid/index.ts
@@ -2,3 +2,4 @@
  * Componentes do grid de disponibilidades
  */
 export { DisponibilidadeGrid } from "./disponibilidade-grid"
+export { GridTurno } from "./grid-turno"
