+++ src/features/disponibilidade-professores/components/index.ts
@@ -1,11 +1,5 @@
 /**
  * Componentes do módulo de Disponibilidade de Professores
  */
-export { CreateDisponibilidadeFormDialog } from "./create-disponibilidade-form-dialog"
-export { EditDisponibilidadeFormDialog } from "./edit-disponibilidade-form-dialog"
-export { DeleteDisponibilidadeAlertDialog } from "./delete-disponibilidade-alert-dialog"
-export { DisponibilidadesDataTable } from "./data-table/disponibilidades-data-table"
-export { disponibilidadeColumns } from "./data-table/disponibilidade-columns"
-export { DisponibilidadeActionRowDropdownMenu } from "./data-table/disponibilidade-action-row-dropdown-menu"
 export { ResumoHorariosCards } from "./resumo-horarios-cards"
 export { DisponibilidadeGrid } from "./grid"
